Added react select library and organizes files into folders
